spiral: split artifacts, uploads, and Lasagna publishing out of spiral_service
spiral_service.py held four unrelated responsibilities in one 3100-line
module. Three of them move out, each with an explicit interface back to the
lifecycle orchestrator rather than free access to ServiceState:

* service_artifacts.py — Artifact/ArtifactRegistry. Self-contained: it knows
  directories, digests and in-flight downloads, and nothing about sessions,
  uploads or previews. The orchestrator supplies the kind/session namespacing
  and the retention policy.

* service_uploads.py — upload transfer, per-kind content validation, and
  publication (ephemeral inputs and the content-addressed store of uploaded
  resume checkpoints). UploadManager reaches the service only through
  UploadEnvironment: where the output directory is, whether a session is
  loaded, that session's ephemeral folder, which checkpoint it resumed from,
  and whether one more ephemeral input may be admitted. It does not see the
  fit session, the status snapshot, the event buffer, the artifact registry
  or the ephemeral ledger; finalize hands back a record and the caller
  decides what to do with it. Duplicate-identity and quota checks stay with
  the orchestrator, which owns that ledger.

* lasagna_publish.py — LasagnaPublisher plus the preview mapping helpers.
  Publication now owns one progress path: the stages it declares, the Lasagna
  job polls, and the one stage only the Lasagna console reports all funnel
  through LasagnaPublisher._report into a single caller-supplied callback.
  The nested start_stage closure, the direct _preview_publish mutations from
  the log relay thread, and the ad-hoc timing prints are gone from the
  service. The publisher's remaining view of the service is four callbacks:
  subprocess attach/detach, session validity, and the previous raw preview
  generation it diffs against.

service_http.py carries the vocabulary all three share (ApiError, safe
relative names, path resolution, chunked hashing) so no extracted module has
to import another for it.

No HTTP surface change: routes, schemas, status codes and API_VERSION 18 are
untouched, and the preview stage names and status fields are byte-identical.
